Fix type errors

+++ model/command/source/command_source_kelvin.py
@@ -1,5 +1,5 @@
-from typing import Iterable, List, Literal
-from pydantic import Field, model_validator
+from typing import Iterable, Literal
+from pydantic import Field
 from .command_source_base import CommandSourceBase
 from ...buffer_types import RgbBuffer
 
@@ -128,6 +128,6 @@ class CommandSourceKelvin(CommandSourceBase):
             12000: (0.765,	0.820,	1),
         }
 
-        kelvin = round(self.kelvin, -2)
+        kelvin = int(round(self.kelvin, -2))
         return kelvin_table.get(kelvin, (0, 0, 0))
